+++ src/api/routers/__init__.py
@@ -1,1 +1,2 @@
-from .user_task import router as user_tasks_router
+from .hello import router as hello_router
+from .user_task import router as user_task_router
