+++ src/main.py
@@ -1,13 +1,15 @@
 from fastapi import FastAPI
 from telegram.ext import ApplicationBuilder, CommandHandler
 
-from src.api.routers import router
+from src.api.routers import hello_router, user_task_router
 from src.bot.handlers import start
 from src.core.settings import settings
 
+
 app = FastAPI()
 
-app.include_router(router)
+app.include_router(hello_router)
+app.include_router(user_task_router)
 
 
 def create_bot():
